refactor: fold the Filesearch resolver into Assistants
"Filesearch" is a dead concept now that the OpenAI filesearch schemas and
context are gone — the resolver only ever did assistant work. Its twelve
functions move into the existing GlificWeb.Resolvers.Assistants (which already
held create_knowledge_base/clone_assistant/list_assistant_config_versions),
consolidating all assistant resolvers into one module named for its
Glific.Assistants context, per the plural-resolver convention.

Also renames the internal Absinthe objects :filesearch_mutations/:filesearch_queries
to :assistant_mutations/:assistant_queries and repoints their import_fields in
schema.ex. These are server-internal names, not part of the API.

The client-facing upload_filesearch_file mutation field is deliberately left
as-is — the frontend queries it, so renaming it would be an API break.

Refs #4933
